requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
asserts==0.12.0
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.4
python-json-logger==2.0.7
requests==2.28.2
urllib3==1.26.18